# project-1/Asgun-alt/pkg/user/controllers/users/httpuser.go
package users

import (
	"errors"
	"net/http"

	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/app/middlewares"
	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/pkg/common/controller"
	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/pkg/domain/request"
	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/pkg/helper/errcode"
	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/pkg/helper/validate"
	"github.com/gnumi34/golang-mentoring/tree/main/project-1/Asgun-alt/pkg/user/service/users"
	"github.com/labstack/echo/v4"
)

var (
	badRequest = http.StatusBadRequest
)

type UserController struct {
	handler controller.BaseController
	usecase users.UsersUsecaseInterface
}

func NewUserController(userUsecase users.UsersUsecaseInterface) *UserController {
	return &UserController{usecase: userUsecase}
}

// Login godoc
// @Summary      Login
// @Description  If user is exists in the database, Generate and RETURN user token.
// @Param 		 User body request.LoginUser true "Login"
// @Tags         Login
// @Accept       json
// @Produce      json
// @Success      200
// @Router       /login [post]
func (uc *UserController) Login(c echo.Context) error {
	req := request.LoginUser{}
	bindErr := c.Bind(&req)
	if bindErr != nil {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("bind data error"))
	}

	ctx := c.Request().Context()
	user, err := uc.usecase.Login(ctx, req.ToLoginDomain())
	if err != nil {
		errCode, errMessage := errcode.ErrorUnauthorizedCheck(err)
		return uc.handler.ErrorResponse(c, errCode, errMessage)
	}

	return uc.handler.SuccessDataResponse(c, map[string]string{"token": user.Token})
}

// Protected route godoc
// @Summary      Protected user route
// @Description  Protected route can only be accessed if the the user has valid JWT token.
// @Param 		 Authorization header string true "Insert your access token" default(Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJVc2VySUQiOiJkMTk3YzY1Ny0zMDY1LTQ0MjYtYmY4ZS05YmJhYWYwYjY5MDciLCJleHAiOjE2NjA4NzQ3MjR9.FV-RV_55zOMO4qD1vjmY0m1ZmeRsvDfMchlbqXjcpkc)
// @Tags         Protected
// @Success      200
// @Router       /user/protected [get]
func (uc *UserController) Protected(c echo.Context) error {
	token, err := middlewares.ValidateAuthorization(c)
	if token == nil || err != nil {
		status, err := errcode.ErrorUnauthorizedCheck(err)
		return uc.handler.ErrorResponse(c, status, err)
	}

	return uc.handler.SuccessMessageResponse(c, "hello this route is protected")
}

// GetUser godoc
// @Summary      Show an account
// @Description  validate username and password, if user is exists in the database RETURN valid user
// @Param 		 User body request.GetUser true "validate user"
// @Tags         GetUser
// @Accept       json
// @Produce      json
// @Success      200
// @Router       /users/get-user [post]
func (uc *UserController) GetUser(c echo.Context) error {
	req := request.GetUser{}
	bindErr := c.Bind(&req)
	if bindErr != nil {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("bind data error"))
	}
	ctx := c.Request().Context()

	userData, err := uc.usecase.GetUser(ctx, req.ToGetUserDomain())
	if err != nil {
		errCode, errMessage := errcode.ErrorGetUserCheck(err)
		return uc.handler.ErrorResponse(c, errCode, errMessage)
	}

	if userData.Username == "" || userData.Password == "" {
		return uc.handler.ErrorResponse(c, http.StatusNotFound, errors.New("invalid user or user has been deleted"))
	}
	return uc.handler.SuccessMessageResponse(c, "valid user")
}

// AddUser godoc
// @Summary      Add user
// @Description  Add new user to the database, ID is generated by the API, password  is saved with BCrypt hash after passess validation.
// @Param 		 User body request.AddUser true "Add User"
// @Tags         Add new user
// @Accept       json
// @Produce      json
// @Success      200  {object}  users.UsersDomain
// @Router       /users/ [post]
func (uc *UserController) AddUser(c echo.Context) error {
	req := request.AddUser{}
	bindErr := c.Bind(&req)
	if bindErr != nil {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("bind data error"))
	}
	ctx := c.Request().Context()

	validErr := validate.Validator(req)
	if validErr != nil {
		return uc.handler.ErrorValidationResponse(c, badRequest, validErr)
	}
	if !validate.ValidatePassword(req.Password_1) {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("password must have 1 uppercase letter, 1 number, 1 special character"))
	}

	_, result := uc.usecase.AddUser(ctx, req.ToUserDomain())
	if result != nil {
		errCode, errMessage := errcode.CheckErrorAddUsers(result)
		return uc.handler.ErrorResponse(c, errCode, errMessage)
	}
	return uc.handler.SuccessOkResponse(c)
}

// UpdateUser godoc
// @Summary      Update User
// @Description  Update the user to the database.
// @Param 		 User body request.UpdateUser true "Update User"
// @Tags         Update User
// @Accept       json
// @Produce      json
// @Success      200  {object}  users.UsersDomain
// @Router       /users/ [put]
func (uc *UserController) UpdateUser(c echo.Context) error {
	req := request.UpdateUser{}
	bindErr := c.Bind(&req)
	if bindErr != nil {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("bind data error"))
	}

	if req.Password_1 != "" && !validate.ValidatePassword(req.Password_1) {
		return uc.handler.ErrorResponse(c, badRequest, errors.New("password must have 1 uppercase letter, 1 number, 1 special character"))
	}

	ctx := c.Request().Context()
	_, result := uc.usecase.UpdateUser(ctx, req.ToUpdateUserDomain())

	if result != nil {
		errCode, errMessage := errcode.ErrorUpdateUsersCheck(result)
		return uc.handler.ErrorResponse(c, errCode, errMessage)
	}
	return uc.handler.SuccessOkResponse(c)

}

// DeleteUser godoc
// @Summary      Delete User
// @Description  Soft delete user by passing the user ID.
// @Param 		 id path string true "delete user"
// @Tags         (Soft) Delete User
// @Accept       json
// @Produce      json
// @Success      200
// @Router       /users/{id} [delete]
func (uc *UserController) DeleteUser(c echo.Context) error {
	id := c.Param("id")

	ctx := c.Request().Context()
	result := uc.usecase.DeleteUser(ctx, id)
	if result != nil {
		errCode, errMessage := errcode.ErrorDeleteUsersCheck(result)
		return uc.handler.ErrorResponse(c, errCode, errMessage)
	}
	return uc.handler.SuccessOkResponse(c)
}